[Refactor] Set proper namespace for NotificationFactory
